requirements.txt
fastapi
uvicorn
python-multipart

pandas
numpy
openpyxl

ydata-profiling

langgraph
langchain
langchain-openai

pydantic
python-dotenv
reportlab

pyarrow
fastparquet